+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
+* Add a timeout to the `load` function of the `StandCloudLoader` class.
+  [[PR-166](https://github.com/everypinio/hardpy/pull/166)]
 * Fix logic for processing spacebar pressing.
   [[PR-164](https://github.com/everypinio/hardpy/pull/164)]
 * Fix the display of the module duration after the operator panel has been restarted.